requirements.txt
langchain==0.3.17
langchain-core==0.3.34
langchain-community==0.3.16
llama-parse==0.5.20
cohere==5.13.1
langchain-cohere==0.3.4
pymilvus==2.5.4
langchain-milvus==0.1.7
groq==0.13.1
langchain-groq==0.2.1
tavily-python==0.5.0
pydantic==2.10.6
gradio==5.17.0
python-dotenv==1.0.1
